Add files via upload
The main role of this project is to carry out the predicting models and analysis techniques to distinguish student's facial conduct during school hours, facial acknowledgment to decide whether the student is sitting at their work area during work hours, and fatigue recognition acknowledgment to decide when the student is drained so that work can be changed in accordance with ideal occasions. Furthermore, the motivation behind the investigation is likewise to survey the different elements that straightforwardly affect student's performance. In light of these elements, teachers and professors can embrace different administration practices and instruments that increase students' satisfaction, which thus can inspire to perform better.